Create BenjaminSterlingSilver_Resume.html
This is an initial draft of my resume being displayed in HTML for a Web Development Foundations course at Western Governors University. All work this was based on, with traces displayed here, is my own work, and will be changed before turning in the project. Additional changes to be implemented before submission are the establishment of a backend (in JavaScript), as well as implemented formatting (utilizing CSS) and more pages added (in the form of a cover letter and page that outlines my career goals).